Move image dataset into separate module
